refactor: decouple action methods from AutoPilotContext to TrainActionManager
