+++ src/main/java/letrain/vehicle/impl/rail/TrainAutoPilotContext.java
@@ -64,66 +64,6 @@ public class TrainAutoPilotContext implements AutoPilotContext {
     }
 
     @Override
-    public void ensureForkRoute(Segment from, Segment to) {
-        if (train.getModel() == null) return;
-        RailwayGraph graph = train.getModel().getRailwayGraph();
-        if (graph == null) return;
-
-        // Find the fork between 'from' and 'to' and set the correct route
-        var fromSteps = from.getSteps();
-        var toSteps = to.getSteps();
-        if (fromSteps == null || toSteps == null) return;
-
-        var f1 = fromSteps.getFirst();
-        var f2 = fromSteps.getSecond();
-        var t1 = toSteps.getFirst();
-        var t2 = toSteps.getSecond();
-
-        RailNode node = null;
-        if (f1 != null && f1.getRailNode() != null) {
-            var n1 = f1.getRailNode();
-            if ((t1 != null && n1.equals(t1.getRailNode())) || (t2 != null && n1.equals(t2.getRailNode()))) {
-                node = n1;
-            }
-        }
-        if (node == null && f2 != null && f2.getRailNode() != null) {
-            var n2 = f2.getRailNode();
-            if ((t1 != null && n2.equals(t1.getRailNode())) || (t2 != null && n2.equals(t2.getRailNode()))) {
-                node = n2;
-            }
-        }
-
-        if (node == null) {
-            org.slf4j.LoggerFactory.getLogger(TrainAutoPilotContext.class)
-                .warn("[FORK] ensureForkRoute {}->{}: no shared node found", from.getId(), to.getId());
-            return;
-        }
-        if (!(node.getTrack() instanceof ForkRailTrack fork)) {
-            org.slf4j.LoggerFactory.getLogger(TrainAutoPilotContext.class)
-                .debug("[FORK] ensureForkRoute {}->{}: shared node is not a fork ({})", from.getId(), to.getId(), node.getTrack());
-            return;
-        }
-
-        // Use the fork node's outSteps directly (getNextSteps goes to wrong node)
-        for (var step : node.getOutSteps()) {
-            Segment nextSeg = graph.getSegment(step);
-            org.slf4j.LoggerFactory.getLogger(TrainAutoPilotContext.class)
-                .debug("[FORK] ensureForkRoute {}->{}: outStep dir={} seg={}", from.getId(), to.getId(), step.getDir(), nextSeg != null ? nextSeg.getId() : "null");
-            if (nextSeg != null && nextSeg.equals(to)) {
-                boolean altNeeded = isAlternativeRouteNeeded(fork, step.getDir());
-                org.slf4j.LoggerFactory.getLogger(TrainAutoPilotContext.class)
-                    .info("[FORK] ensureForkRoute {}->{}: MATCH fork={} altNeeded={} currentAlt={}", from.getId(), to.getId(), fork.getId(), altNeeded, fork.isUsingAlternativeRoute());
-                if (fork.isUsingAlternativeRoute() != altNeeded) {
-                    fork.flipRoute();
-                }
-                return;
-            }
-        }
-        org.slf4j.LoggerFactory.getLogger(TrainAutoPilotContext.class)
-            .warn("[FORK] ensureForkRoute {}->{}: no outStep leads to target seg", from.getId(), to.getId());
-    }
-
-    @Override
     public boolean isAtTarget(Waypoint wp) {
         if (train.getModel() == null) return false;
         switch (wp.type()) {
@@ -161,8 +101,6 @@ public class TrainAutoPilotContext implements AutoPilotContext {
         return false;
     }
 
-
-
     private Point findTargetPosition(Waypoint wp) {
         if (train.getModel() == null) return null;
         switch (wp.type()) {
@@ -174,21 +112,5 @@ public class TrainAutoPilotContext implements AutoPilotContext {
                 return sensor != null ? sensor.getPosition() : null;
         }
         return null;
-    }
-
-    @Override
-    public void forceSegmentReset() {
-        train.forceSegmentReset();
-    }
-
-    @Override
-    public void notifySegmentOccupied(Segment segment) {
-        train.notifySegmentOccupied(segment);
-    }
-
-    private boolean isAlternativeRouteNeeded(ForkRailTrack fork, letrain.map.Dir targetDir) {
-        // Check if the targetDir is the alternative route of the fork
-        var alt = fork.getRouter().getAlternativeRoute();
-        return alt != null && alt.getValue() == targetDir;
     }
 }
